typo in "Decentralization" trends>tends
This is a principle of the design of distributed systems, including societies. It points out that any single common point which is involved in any operation trends to limit the way the system scales, and produce a single point of complete failure.

Should be 
This is a principle of the design of distributed systems, including societies. It points out that any single common point which is involved in any operation `tends` to limit the way the system scales, and produce a single point of complete failure.