Typo in Principle of Least Power : "uses">"users"
In principle 7, paragraph 2 : 
The choice of language is a common design choice. The low power end of the scale is typically simpler to design, implement and use, but the high power end of the scale has all the attraction of being an open-ended hook into which anything can be placed, a door to uses bounded only by the imagination of the programmer.

should be :
The choice of language is a common design choice. The low power end of the scale is typically simpler to design, implement and use, but the high power end of the scale has all the attraction of being an open-ended hook into which anything can be placed, a door to `users` bounded only by the imagination of the programmer.